+++ __tests__/rules/substituteSpaceAtRules.test.js
@@ -186,6 +186,32 @@ it('parses @space! only for requested bp', () => {
   })
 })
 
+it('parses @space with px', () => {
+  const input = `
+    body article .test {
+      @space margin-top 40px desktop;
+      font-size: 18px;
+    }
+  `
+
+  const output = `
+    body article .test {
+      font-size: 18px;
+    }
+
+    @media (min-width: 1024px){
+      body article .test {
+        margin-top: 40px;
+      }
+    }
+  `
+
+  return run(input, DEFAULT_CFG).then(result => {
+    expect(result.css).toMatchCSS(output)
+    expect(result.warnings().length).toBe(0)
+  })
+})
+
 it('parses @space for theme(..)', () => {
   const input = `
     body article .test {